Compound class for OSP snapshot compounds

An R6 class that represents a compound in an OSP snapshot. This class provides methods to access different properties of a compound and display a summary of its information.

Public fields

data

The raw data of the compound

Active bindings

name

The name of the compound

is_small_molecule

Whether the compound is a small molecule

plasma_protein_binding_partner

The plasma protein binding partner of the compound

molecular_weight

The molecular weight of the compound

molecular_weight_unit

The unit of the molecular weight

lipophilicity

The lipophilicity data of the compound

fraction_unbound

The fraction unbound data of the compound

solubility

The solubility data of the compound

intestinal_permeability

The intestinal permeability data of the compound

pka_types

The pKa types of the compound

processes

The processes of the compound

calculation_methods

The calculation methods of the compound

parameters

The parameters of the compound

Methods

Public methods

Compound$new()

Compound$print()

Compound$clone()

Method new()

Create a new Compound object

Usage

Compound$new(data)

Arguments

data

Raw compound data from a snapshot

Returns

A new Compound object

Method print()

Print a summary of the compound

Usage

Compound$print(...)

Arguments

...

Additional arguments passed to print methods

Returns

Invisibly returns the object

Method clone()

The objects of this class are cloneable with this method.

Usage

Compound$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.